requirements.txt
# requirements.txt
# 核心依赖
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7

# NLP和向量化
sentence-transformers==2.2.2
transformers==4.35.0
torch>=2.0.0

# 向量数据库
faiss-cpu==1.7.4  # 或者 faiss-gpu 如果有GPU

# 文档处理
PyPDF2==3.0.1
python-docx==1.1.0
chardet==5.2.0

# 工具库
numpy==1.24.3
scikit-learn==1.3.0
pyyaml==6.0.1
jieba==0.42.1
tiktoken==0.5.1

# 可选：如果使用OpenAI
openai==1.3.0

# 开发工具
pytest==7.4.3
black==23.10.1
pylint==3.0.2